Extract LocalRestApiSettingTab into its own file
Move the settings tab class from main.ts into src/settings.ts to reduce
file size and improve separation of concerns. Use a type-only import in
settings.ts to avoid circular dependency.

Also clean up unused imports in main.ts that were only needed by the
settings tab class, add definite assignment assertions on properties
initialized in onload(), and fix tsconfig.json deprecation warnings by
removing unused baseUrl and switching moduleResolution to 'bundler'.
